优化微信返回值
